+++ Sources/Swift/Integrations/SessionReplay/SentryReplayOptions.swift
@@ -16,7 +16,6 @@ public class SentryReplayOptions: NSObject, SentryRedactOptions {
         public static let enableViewRendererV2: Bool = true
         public static let enableFastViewRendering: Bool = false
         public static let quality: SentryReplayQuality = .medium
-        public static let disableInDangerousEnvironment: Bool = true
 
         // The following properties are public because they are used by SentrySwiftUI.
 
@@ -217,17 +216,6 @@ public class SentryReplayOptions: NSObject, SentryRedactOptions {
     public var enableFastViewRendering: Bool
 
     /**
-     * Due to internal changes with the release of Liquid Glass on iOS 26.0, the masking of text and images can not be reliably guaranteed.
-     
-     * Therefore the session replay integration is disabled by default unless the environment is detected as safe.
-     *
-     * - Important: This flag allows to re-enable the session replay integration on iOS 26.0 and later, but please be aware that text and images may not be masked as expected.
-     *
-     * - Note: See [GitHub issues #6389](https://github.com/getsentry/sentry-cocoa/issues/6389) for more information.
-     */
-    public var disableInDangerousEnvironment: Bool
-
-    /**
      * Defines the quality of the session replay.
      *
      * Higher bit rates better quality, but also bigger files to transfer.
@@ -302,7 +290,6 @@ public class SentryReplayOptions: NSObject, SentryRedactOptions {
             maskAllImages: nil,
             enableViewRendererV2: nil,
             enableFastViewRendering: nil,
-            disableInDangerousEnvironment: nil,
             maskedViewClasses: nil,
             unmaskedViewClasses: nil,
             quality: nil,
@@ -332,7 +319,6 @@ public class SentryReplayOptions: NSObject, SentryRedactOptions {
             enableViewRendererV2: (dictionary["enableViewRendererV2"] as? NSNumber)?.boolValue
             ?? (dictionary["enableExperimentalViewRenderer"] as? NSNumber)?.boolValue,
             enableFastViewRendering: (dictionary["enableFastViewRendering"] as? NSNumber)?.boolValue,
-            disableInDangerousEnvironment: (dictionary["disableInDangerousEnvironment"] as? NSNumber)?.boolValue,
             maskedViewClasses: (dictionary["maskedViewClasses"] as? NSArray)?.compactMap({ element in
                 NSClassFromString((element as? String) ?? "")
             }),
@@ -367,8 +353,7 @@ public class SentryReplayOptions: NSObject, SentryRedactOptions {
         maskAllText: Bool = DefaultValues.maskAllText,
         maskAllImages: Bool = DefaultValues.maskAllImages,
         enableViewRendererV2: Bool = DefaultValues.enableViewRendererV2,
-        enableFastViewRendering: Bool = DefaultValues.enableFastViewRendering,
-        disableInDangerousEnvironment: Bool = DefaultValues.disableInDangerousEnvironment
+        enableFastViewRendering: Bool = DefaultValues.enableFastViewRendering
     ) {
         // - This initializer is publicly available for Swift, but not for Objective-C, because automatically bridged Swift initializers
         //   with default values result in a single initializer requiring all parameters.
@@ -383,7 +368,6 @@ public class SentryReplayOptions: NSObject, SentryRedactOptions {
             maskAllImages: maskAllImages,
             enableViewRendererV2: enableViewRendererV2,
             enableFastViewRendering: enableFastViewRendering,
-            disableInDangerousEnvironment: disableInDangerousEnvironment,
             maskedViewClasses: nil,
             unmaskedViewClasses: nil,
             quality: nil,
@@ -403,7 +387,6 @@ public class SentryReplayOptions: NSObject, SentryRedactOptions {
         maskAllImages: Bool?,
         enableViewRendererV2: Bool?,
         enableFastViewRendering: Bool?,
-        disableInDangerousEnvironment: Bool?,
         maskedViewClasses: [AnyClass]?,
         unmaskedViewClasses: [AnyClass]?,
         quality: SentryReplayQuality?,
@@ -419,7 +402,6 @@ public class SentryReplayOptions: NSObject, SentryRedactOptions {
         self.maskAllImages = maskAllImages ?? DefaultValues.maskAllImages
         self.enableViewRendererV2 = enableViewRendererV2 ?? DefaultValues.enableViewRendererV2
         self.enableFastViewRendering = enableFastViewRendering ?? DefaultValues.enableFastViewRendering
-        self.disableInDangerousEnvironment = disableInDangerousEnvironment ?? DefaultValues.disableInDangerousEnvironment
         self.maskedViewClasses = maskedViewClasses ?? DefaultValues.maskedViewClasses
         self.unmaskedViewClasses = unmaskedViewClasses ?? DefaultValues.unmaskedViewClasses
         self.quality = quality ?? DefaultValues.quality